Updated to libdwarf-20070223 Some formatting was a bit off because of macros without terminating semicolons. Changed all the libdwarf/dwarf_util.h macros so now all uses require terminating semicolons and re-indented libdwarf. Fixed an output bug in dwarfdump frame output (was missing a newline in some cases).

+++ dwarfdump/print_reloc.c
@@ -385,9 +385,9 @@ print_reloc_information_64(int section_no, Dwarf_Small * buf,
 	/* This works for the Elf64_Rel in linux */
 	Elf64_Rel *p = (Elf64_Rel *) (buf + off);
 
-	printf("%5lu\t<%3d> %-34s%s\n",
+	printf("%5lu\t<%3ld> %-34s%s\n",
 	       (unsigned long int) (p->r_offset),
-	       ELF64_R_SYM(p->r_info),
+	       (long)ELF64_R_SYM(p->r_info),
 	       sym_data[ELF64_R_SYM(p->r_info) - 1].name,
 	       get_reloc_type_names(ELF64_R_TYPE(p->r_info)));
 #else
@@ -398,7 +398,7 @@ print_reloc_information_64(int section_no, Dwarf_Small * buf,
 
 	printf("%5llu\t<%3d> %-34s%s\n",
 	       (unsigned long long int) (p->r_offset),
-	       p->r_sym, sym_data_64[p->r_sym - 1].name,
+	       (long)p->r_sym, sym_data_64[p->r_sym - 1].name,
 	       get_reloc_type_names(p->r_type));
 #endif
     }
